+++ R/checks.R
@@ -104,7 +104,7 @@ check_salinity <- function(x, interval = 15, site, sites) {
 
   # Special case where salinity was bad and not used for calibration
   if (all(is.na(x)))
-    return(rep("Sm:", length(x)))
+    return(rep("", length(x)))
 
 
   # Jumps